refactor: moved `daide_visitor.py` to `visitor.py`
