+++ src/daidepp/__init__.py
@@ -1,6 +1,6 @@
-from daidepp.daide_visitor import *
 from daidepp.grammar import *
 from daidepp.keywords import *
+from daidepp.visitor import *
 
 try:
     from importlib.metadata import version
