+++ src/daidepp/visitor.py
